Store per-peak sum waveform Also restore sane computations for height and index_of_maximum

+++ pax/plugins/signal_processing/SumWaveform.py
@@ -19,7 +19,7 @@ class SumWaveform(plugin.TransformPlugin):
         for postfix in ('', '_raw'):
             for detector, chs in self.config['channels_in_detector'].items():
                 event.sum_waveforms.append(datastructure.SumWaveform(
-                    samples=np.zeros(event.length()),
+                    samples=np.zeros(event.length(), dtype=np.float32),
                     name=detector + postfix,
                     channel_list=np.array(list(chs), dtype=np.uint16),
                     detector=detector
@@ -39,7 +39,7 @@ class SumWaveform(plugin.TransformPlugin):
             else:
                 baseline_to_subtract = self.config['digitizer_reference_baseline'] - pulse.baseline
 
-            w = baseline_to_subtract - pulse.raw_data.astype(np.float64)
+            w = baseline_to_subtract - pulse.raw_data.astype(np.float32)
 
             adc_to_pe = utils.adc_to_pe(self.config, pulse.channel)
 
@@ -61,7 +61,7 @@ class SumWaveform(plugin.TransformPlugin):
             left_in_pulse = hit.left - pulse.left
             right_in_pulse = hit.right - pulse.left
             w = (self.config['digitizer_reference_baseline'] - pulse.baseline) - \
-                pulse.raw_data[left_in_pulse:right_in_pulse+1].astype(np.float64)
+                pulse.raw_data[left_in_pulse:right_in_pulse+1].astype(np.float32)
 
             adc_to_pe = utils.adc_to_pe(self.config, hit.channel)
